requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-marshmallow
marshmallow-sqlalchemy
pandas
openpyxl
requests
m3u8
celery==5.4.0
redis==5.0.1
eventlet
python-dotenv
psycopg2-binary
sqlalchemy-utils
flask-apscheduler
flask-login
psutil
gunicorn
yt-dlp
Flask-Session
PyJWT
numpy<2.0.0
cloudscraper
playwright
playwright-stealth
beautifulsoup4